Use size_t instead of uint64_t for array dimensions and loop indices in padding.h

+++ include/padding.h
@@ -16,31 +16,31 @@ namespace tomocam {
 
         if ((int)std::round(factor) == 1) { return arr.clone(); }
 
-        uint64_t n3 = static_cast<uint64_t>(factor * arr.ncols());
+        size_t n3 = static_cast<size_t>(factor * arr.ncols());
         if (n3 % 2 == 0) { n3 -= 1; }
-        uint64_t pad_size = n3 - arr.ncols();
+        size_t pad_size = n3 - arr.ncols();
 
         // allocate return array
         dims_t dims = {arr.nslices(), arr.nrows(), n3};
         Array<T> arr2(dims);
 
         // assume symmetric padding by default
-        uint64_t d = pad_size / 2;
+        size_t d = pad_size / 2;
         if (pad_type == PadType::RIGHT) {
             d = 0;
         } else if (pad_type == PadType::LEFT) {
             d = pad_size;
         }
 
-        for (uint64_t i = 0; i < dims.x(); i++)
-            for (uint64_t j = 0; j < dims.y(); j++)
-                for (uint64_t k = 0; k < dims.z(); k++)
+        for (size_t i = 0; i < dims.x(); i++)
+            for (size_t j = 0; j < dims.y(); j++)
+                for (size_t k = 0; k < dims.z(); k++)
                     arr2[{i, j, k + d}] = arr[{i, j, k}];
         return arr2;
     }
 
     template <typename T>
-    auto crop1d(const Array<T> &arr, uint64_t crop_size,
+    auto crop1d(const Array<T> &arr, size_t crop_size,
         PadType pad_type) -> Array<T> {
 
         if (crop_size == 0) { return arr.clone(); }
@@ -55,9 +55,9 @@ namespace tomocam {
             d = crop_size;
         }
 
-        for (uint64_t i = 0; i < dims.x(); i++)
-            for (uint64_t j = 0; j < dims.y(); j++)
-                for (uint64_t k = 0; k < dims.z(); k++)
+        for (size_t i = 0; i < dims.x(); i++)
+            for (size_t j = 0; j < dims.y(); j++)
+                for (size_t k = 0; k < dims.z(); k++)
                     arr2[{i, j, k}] = arr[{i, j, k + d}];
         return arr2;
     }
@@ -66,9 +66,9 @@ namespace tomocam {
     auto pad2d(const Array<T> &arr, T factor, PadType pad_type) -> Array<T> {
 
         if ((int)std::round(factor) == 1) { return arr.clone(); }
-        auto n2 = static_cast<uint64_t>(factor * arr.nrows());
+        auto n2 = static_cast<size_t>(factor * arr.nrows());
         if (n2 % 2 == 0) { n2 -= 1; }
-        auto n3 = static_cast<uint64_t>(factor * arr.ncols());
+        auto n3 = static_cast<size_t>(factor * arr.ncols());
         if (n3 % 2 == 0) { n3 -= 1; }
 
         // create and initialize return array
@@ -84,9 +84,9 @@ namespace tomocam {
             d = pad_size;
         }
 
-        for (uint64_t i = 0; i < dims.x(); i++) {
-            for (uint64_t j = 0; j < dims.y(); j++) {
-                for (uint64_t k = 0; k < dims.z(); k++) {
+        for (size_t i = 0; i < dims.x(); i++) {
+            for (size_t j = 0; j < dims.y(); j++) {
+                for (size_t k = 0; k < dims.z(); k++) {
                     arr2[{i, j + d.y(), k + d.z()}] = arr[{i, j, k}];
                 }
             }
@@ -114,9 +114,9 @@ namespace tomocam {
             d = crop_size;
         }
 
-        for (uint64_t i = 0; i < new_dims.x(); i++) {
-            for (uint64_t j = 0; j < new_dims.y(); j++) {
-                for (uint64_t k = 0; k < new_dims.z(); k++) {
+        for (size_t i = 0; i < new_dims.x(); i++) {
+            for (size_t j = 0; j < new_dims.y(); j++) {
+                for (size_t k = 0; k < new_dims.z(); k++) {
                     arr2[{i, j, k}] = arr[{i, j + d.y(), k + d.z()}];
                 }
             }
@@ -130,11 +130,11 @@ namespace tomocam {
         if ((int)std::round(factor) == 1) { return arr.clone(); }
 
         // calculate new dims
-        auto n1 = static_cast<uint64_t>(static_cast<T>(arr.nslices() * factor));
+        auto n1 = static_cast<size_t>(static_cast<T>(arr.nslices() * factor));
         if (n1 % 2 == 0) { n1 -= 1; }
-        auto n2 = static_cast<uint64_t>(static_cast<T>(arr.nrows() * factor));
+        auto n2 = static_cast<size_t>(static_cast<T>(arr.nrows() * factor));
         if (n2 % 2 == 0) { n2 -= 1; }
-        auto n3 = static_cast<uint64_t>(static_cast<T>(arr.ncols() * factor));
+        auto n3 = static_cast<size_t>(static_cast<T>(arr.ncols() * factor));
         if (n3 % 2 == 0) { n3 -= 1; }
 
         // allocate return array
@@ -151,9 +151,9 @@ namespace tomocam {
             d = pad_size;
         }
 
-        for (uint64_t i = 0; i < dims.x(); i++) {
-            for (uint64_t j = 0; j < dims.y(); j++) {
-                for (uint64_t k = 0; k < dims.z(); k++) {
+        for (size_t i = 0; i < dims.x(); i++) {
+            for (size_t j = 0; j < dims.y(); j++) {
+                for (size_t k = 0; k < dims.z(); k++) {
                     arr2[{i + d.x(), j + d.y(), k + d.z()}] = arr[{i, j, k}];
                 }
             }
@@ -177,9 +177,9 @@ namespace tomocam {
         else if (pad_type == PadType::RIGHT)
             d = crop_size;
 
-        for (uint64_t i = 0; i < new_dims.x(); i++) {
-            for (uint64_t j = 0; j < new_dims.y(); j++) {
-                for (uint64_t k = 0; k < new_dims.z(); k++) {
+        for (size_t i = 0; i < new_dims.x(); i++) {
+            for (size_t j = 0; j < new_dims.y(); j++) {
+                for (size_t k = 0; k < new_dims.z(); k++) {
                     arr2[{i, j, k}] = arr[{i + d.x(), j + d.y(), k + d.z()}];
                 }
             }
